+++ src/ebfm/coupling/__init__.py
@@ -5,7 +5,7 @@
 from .couplers.helpers import coupling_supported
 from .couplers.helpers import coupling_supported_import_error  # noqa: F401
 
-from .couplers import Coupler, DummyCoupler  # noqa: F401
+from .couplers import Coupler, DummyCoupler, FakeCoupler  # noqa: F401
 
 if coupling_supported:
     from .couplers.yacCoupler import YACCoupler  # noqa: F401